+++ src/main/kotlin/com/example/newsletters/service/RequestDestinationService.kt
@@ -12,9 +12,9 @@ class RequestDestinationService(val repository: RequestDestinationRepository) {
     fun getAll(paging: Pageable) = repository.findAll(paging).map { it.requestDestinationDTO }
     fun getByName(name: String) = repository.findByNameContainingIgnoreCase(name).map { it.requestDestinationDTO }
     fun getByName(name: String, paging: Pageable) = repository.findByNameContainingIgnoreCase(name, paging).map { it.requestDestinationDTO }
-    fun getById(id: Int) = repository.findById(id).map { it.requestDestinationDTO }.orElse(null)
-    fun getByIds(ids: List<Int>) = repository.findAllById(ids).map { it.requestDestinationDTO }
-    fun delete(id: Int) = repository.deleteById(id)
+    fun getById(id: Long) = repository.findById(id).map { it.requestDestinationDTO }.orElse(null)
+    fun getByIds(ids: List<Long>) = repository.findAllById(ids).map { it.requestDestinationDTO }
+    fun delete(id: Long) = repository.deleteById(id)
     fun create(data: RequestDestinationDto) = repository.save(data.requestDestination)
     fun update(data: RequestDestinationDto) = repository.save(data.requestDestination)
 }
